Added contract docs to docusaurus docs
Figured out a way to combine docs rather than having do doc instances.
Ideally this could be automated in the future, but for now, it is a
manual process that isn't too painful. I describe the process in the
README.md. There are some additional fixes in the docs that need to be
made so that certain md files render correctly.